Updated to WordPress 3.3

+++ wp-admin/includes/theme.php
@@ -16,10 +16,19 @@
 function current_theme_info() {
 	$themes = get_themes();
 	$current_theme = get_current_theme();
+
+	if ( ! $themes ) {
+		$ct = new stdClass;
+		$ct->name = $current_theme;
+		return $ct;
+	}
+
 	if ( ! isset( $themes[$current_theme] ) ) {
 		delete_option( 'current_theme' );
 		$current_theme = get_current_theme();
 	}
+
+	$ct = new stdClass;
 	$ct->name = $current_theme;
 	$ct->title = $themes[$current_theme]['Title'];
 	$ct->version = $themes[$current_theme]['Version'];
